split client key handling
