+++ update_client.go
@@ -2,11 +2,8 @@ package main
 
 import (
 	"fmt"
-	"sort"
 	"strings"
-	"time"
 
-	"github.com/atotto/clipboard"
 	"github.com/charmbracelet/bubbles/list"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
@@ -66,433 +63,6 @@ func (m *model) ensureTopicVisible() {
 	} else if b.yPos+b.height > m.topics.vp.YOffset+m.topics.vp.Height {
 		m.topics.vp.SetYOffset(b.yPos + b.height - m.topics.vp.Height)
 	}
-}
-
-// handleClientKey processes keyboard events in client mode.
-func (m *model) handleClientKey(msg tea.KeyMsg) tea.Cmd {
-	var cmds []tea.Cmd
-	switch msg.String() {
-	case "ctrl+d":
-		m.saveCurrent()
-		m.savePlannedTraces()
-		return tea.Quit
-	case "ctrl+c":
-		var idxs []int
-		for i, it := range m.history.items {
-			if it.isSelected != nil && *it.isSelected {
-				idxs = append(idxs, i)
-			}
-		}
-		if len(idxs) > 0 {
-			sort.Ints(idxs)
-			var parts []string
-			items := m.history.list.Items()
-			for _, i := range idxs {
-				if i >= 0 && i < len(items) {
-					hi := items[i].(historyItem)
-					txt := hi.payload
-					if hi.kind != "log" {
-						txt = fmt.Sprintf("%s: %s", hi.topic, hi.payload)
-					}
-					parts = append(parts, txt)
-				}
-			}
-			if err := clipboard.WriteAll(strings.Join(parts, "\n")); err != nil {
-				m.appendHistory("", err.Error(), "log", err.Error())
-			}
-		} else if len(m.history.list.Items()) > 0 {
-			idx := m.history.list.Index()
-			if idx >= 0 {
-				hi := m.history.list.Items()[idx].(historyItem)
-				text := hi.payload
-				if hi.kind != "log" {
-					text = fmt.Sprintf("%s: %s", hi.topic, hi.payload)
-				}
-				if err := clipboard.WriteAll(text); err != nil {
-					m.appendHistory("", err.Error(), "log", err.Error())
-				}
-			}
-		}
-	case "ctrl+x":
-		if m.mqttClient != nil {
-			m.mqttClient.Disconnect()
-			m.connections.manager.Statuses[m.connections.active] = "disconnected"
-			m.connections.manager.Errors[m.connections.active] = ""
-			m.refreshConnectionItems()
-			m.connections.connection = ""
-			m.connections.active = ""
-			m.mqttClient = nil
-		}
-	case "/":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory {
-			cmd := m.startHistoryFilter()
-			if cmd != nil {
-				cmds = append(cmds, cmd)
-			}
-		}
-	case "ctrl+f":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory {
-			m.history.filterQuery = ""
-			m.history.list.FilterInput.SetValue("")
-			m.history.list.SetFilterState(list.Unfiltered)
-			var msgs []Message
-			if m.history.showArchived {
-				msgs = m.history.store.SearchArchived(nil, time.Time{}, time.Time{}, "")
-			} else {
-				msgs = m.history.store.Search(nil, time.Time{}, time.Time{}, "")
-			}
-			m.history.items = make([]historyItem, len(msgs))
-			items := make([]list.Item, len(msgs))
-			for i, mm := range msgs {
-				hi := historyItem{timestamp: mm.Timestamp, topic: mm.Topic, payload: mm.Payload, kind: mm.Kind, archived: mm.Archived}
-				m.history.items[i] = hi
-				items[i] = hi
-			}
-			m.history.list.SetItems(items)
-		}
-	case "space":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory && !m.history.showArchived {
-			idx := m.history.list.Index()
-			if idx >= 0 && idx < len(m.history.items) {
-				if m.history.items[idx].isSelected != nil && *m.history.items[idx].isSelected {
-					m.history.items[idx].isSelected = nil
-				} else {
-					v := true
-					m.history.items[idx].isSelected = &v
-				}
-				m.history.selectionAnchor = idx
-			}
-		}
-	case "shift+up":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory && !m.history.showArchived {
-			if m.history.selectionAnchor == -1 {
-				m.history.selectionAnchor = m.history.list.Index()
-				if m.history.selectionAnchor >= 0 && m.history.selectionAnchor < len(m.history.items) {
-					v := true
-					m.history.items[m.history.selectionAnchor].isSelected = &v
-				}
-			}
-			if m.history.list.Index() > 0 {
-				m.history.list.CursorUp()
-				idx := m.history.list.Index()
-				m.updateSelectionRange(idx)
-			}
-		}
-	case "shift+down":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory && !m.history.showArchived {
-			if m.history.selectionAnchor == -1 {
-				m.history.selectionAnchor = m.history.list.Index()
-				if m.history.selectionAnchor >= 0 && m.history.selectionAnchor < len(m.history.items) {
-					v := true
-					m.history.items[m.history.selectionAnchor].isSelected = &v
-				}
-			}
-			if m.history.list.Index() < len(m.history.list.Items())-1 {
-				m.history.list.CursorDown()
-				idx := m.history.list.Index()
-				m.updateSelectionRange(idx)
-			}
-		}
-	case "tab":
-		if len(m.ui.focusOrder) > 0 {
-			m.focus.Next()
-			m.ui.focusIndex = m.focus.Index()
-			id := m.ui.focusOrder[m.ui.focusIndex]
-			m.setFocus(id)
-			if id == idTopics {
-				if len(m.topics.items) > 0 {
-					m.topics.selected = 0
-					m.ensureTopicVisible()
-				} else {
-					m.topics.selected = -1
-				}
-			}
-		}
-	case "shift+tab":
-		if len(m.ui.focusOrder) > 0 {
-			m.focus.Prev()
-			m.ui.focusIndex = m.focus.Index()
-			id := m.ui.focusOrder[m.ui.focusIndex]
-			m.setFocus(id)
-			if id == idTopics {
-				if len(m.topics.items) > 0 {
-					m.topics.selected = 0
-					m.ensureTopicVisible()
-				} else {
-					m.topics.selected = -1
-				}
-			}
-		}
-	case "left":
-		if m.ui.focusOrder[m.ui.focusIndex] == idTopics && len(m.topics.items) > 0 {
-			m.topics.selected = (m.topics.selected - 1 + len(m.topics.items)) % len(m.topics.items)
-			m.ensureTopicVisible()
-		}
-	case "right":
-		if m.ui.focusOrder[m.ui.focusIndex] == idTopics && len(m.topics.items) > 0 {
-			m.topics.selected = (m.topics.selected + 1) % len(m.topics.items)
-			m.ensureTopicVisible()
-		}
-	case "ctrl+shift+up":
-		id := m.ui.focusOrder[m.ui.focusIndex]
-		if id == idMessage {
-			if m.layout.message.height > 1 {
-				m.layout.message.height--
-				m.message.input.SetHeight(m.layout.message.height)
-			}
-		} else if id == idHistory {
-			if m.layout.history.height > 1 {
-				m.layout.history.height--
-				m.history.list.SetSize(m.ui.width-4, m.layout.history.height)
-			}
-		} else if id == idTopics {
-			if m.layout.topics.height > 1 {
-				m.layout.topics.height--
-			}
-		}
-	case "ctrl+shift+down":
-		id := m.ui.focusOrder[m.ui.focusIndex]
-		if id == idMessage {
-			m.layout.message.height++
-			m.message.input.SetHeight(m.layout.message.height)
-		} else if id == idHistory {
-			m.layout.history.height++
-			m.history.list.SetSize(m.ui.width-4, m.layout.history.height)
-		} else if id == idTopics {
-			m.layout.topics.height++
-		}
-	case "ctrl+a":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory && !m.history.showArchived {
-			allSelected := true
-			for i := range m.history.items {
-				if m.history.items[i].isSelected == nil || !*m.history.items[i].isSelected {
-					allSelected = false
-					break
-				}
-			}
-			if allSelected {
-				for i := range m.history.items {
-					m.history.items[i].isSelected = nil
-				}
-				m.history.selectionAnchor = -1
-			} else {
-				for i := range m.history.items {
-					v := true
-					m.history.items[i].isSelected = &v
-				}
-				if len(m.history.items) > 0 {
-					m.history.selectionAnchor = 0
-				}
-			}
-		}
-	case "ctrl+l":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory && m.history.store != nil {
-			m.history.showArchived = !m.history.showArchived
-			var msgs []Message
-			if m.history.showArchived {
-				msgs = m.history.store.SearchArchived(nil, time.Time{}, time.Time{}, "")
-			} else {
-				msgs = m.history.store.Search(nil, time.Time{}, time.Time{}, "")
-			}
-			m.history.items = make([]historyItem, len(msgs))
-			items := make([]list.Item, len(msgs))
-			for i, mm := range msgs {
-				hi := historyItem{timestamp: mm.Timestamp, topic: mm.Topic, payload: mm.Payload, kind: mm.Kind, archived: mm.Archived}
-				m.history.items[i] = hi
-				items[i] = hi
-			}
-			m.history.list.SetItems(items)
-			if len(items) > 0 {
-				m.history.list.Select(len(items) - 1)
-			} else {
-				m.history.list.Select(-1)
-			}
-		}
-	case "up", "down", "k", "j":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory {
-			// keep current selection and anchor
-		} else if m.ui.focusOrder[m.ui.focusIndex] == idTopics {
-			delta := -1
-			if msg.String() == "down" || msg.String() == "j" {
-				delta = 1
-			}
-			m.scrollTopics(delta)
-		}
-	case "ctrl+s", "ctrl+enter":
-		if m.ui.focusOrder[m.ui.focusIndex] == idMessage {
-			payload := m.message.input.Value()
-			for _, t := range m.topics.items {
-				if t.subscribed {
-					m.message.payloads = append(m.message.payloads, payloadItem{topic: t.title, payload: payload})
-					m.appendHistory(t.title, payload, "pub", fmt.Sprintf("Published to %s: %s", t.title, payload))
-					if m.mqttClient != nil {
-						m.mqttClient.Publish(t.title, 0, false, payload)
-					}
-				}
-			}
-		}
-	case "enter":
-		if m.ui.focusOrder[m.ui.focusIndex] == idTopic {
-			topic := strings.TrimSpace(m.topics.input.Value())
-			if topic != "" && !m.hasTopic(topic) {
-				m.topics.items = append(m.topics.items, topicItem{title: topic, subscribed: true})
-				m.sortTopics()
-				if m.currentMode() == modeTopics {
-					m.rebuildActiveTopicList()
-				}
-				if m.mqttClient != nil {
-					m.mqttClient.Subscribe(topic, 0, nil)
-				}
-				m.appendHistory(topic, "", "log", fmt.Sprintf("Subscribed to topic: %s", topic))
-				m.topics.input.SetValue("")
-			}
-		} else if m.ui.focusOrder[m.ui.focusIndex] == idTopics && m.topics.selected >= 0 && m.topics.selected < len(m.topics.items) {
-			m.toggleTopic(m.topics.selected)
-			m.ensureTopicVisible()
-			if m.currentMode() == modeTopics {
-				m.rebuildActiveTopicList()
-			}
-		}
-	case "a":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory && !m.history.showArchived {
-			if len(m.history.items) == 0 {
-				break
-			}
-			archived := false
-			for i := len(m.history.items) - 1; i >= 0; i-- {
-				it := m.history.items[i]
-				if it.isSelected != nil && *it.isSelected {
-					key := fmt.Sprintf("%s/%020d", it.topic, it.timestamp.UnixNano())
-					if m.history.store != nil {
-						_ = m.history.store.Archive(key)
-					}
-					m.history.items = append(m.history.items[:i], m.history.items[i+1:]...)
-					archived = true
-				}
-			}
-			if !archived {
-				idx := m.history.list.Index()
-				if idx >= 0 && idx < len(m.history.items) {
-					it := m.history.items[idx]
-					key := fmt.Sprintf("%s/%020d", it.topic, it.timestamp.UnixNano())
-					if m.history.store != nil {
-						_ = m.history.store.Archive(key)
-					}
-					m.history.items = append(m.history.items[:idx], m.history.items[idx+1:]...)
-				}
-			}
-			items := make([]list.Item, len(m.history.items))
-			for i, it := range m.history.items {
-				it.isSelected = nil
-				m.history.items[i] = it
-				items[i] = it
-			}
-			m.history.list.SetItems(items)
-			if len(m.history.items) == 0 {
-				m.history.list.Select(-1)
-			} else if m.history.list.Index() >= len(m.history.items) {
-				m.history.list.Select(len(m.history.items) - 1)
-			}
-			m.history.selectionAnchor = -1
-		}
-	case "delete":
-		if m.ui.focusOrder[m.ui.focusIndex] == idHistory && !m.history.showArchived {
-			if len(m.history.items) == 0 {
-				break
-			}
-			hasSelection := false
-			for i := range m.history.items {
-				if m.history.items[i].isSelected != nil && *m.history.items[i].isSelected {
-					v := true
-					m.history.items[i].isMarkedForDeletion = &v
-					hasSelection = true
-				}
-			}
-			if !hasSelection {
-				idx := m.history.list.Index()
-				if idx >= 0 && idx < len(m.history.items) {
-					v := true
-					m.history.items[idx].isMarkedForDeletion = &v
-				}
-			}
-			m.confirmReturnFocus = m.ui.focusOrder[m.ui.focusIndex]
-			m.startConfirm("Delete selected messages? [y/n]", "", func() {
-				for i := len(m.history.items) - 1; i >= 0; i-- {
-					it := m.history.items[i]
-					if it.isMarkedForDeletion != nil && *it.isMarkedForDeletion {
-						key := fmt.Sprintf("%s/%020d", it.topic, it.timestamp.UnixNano())
-						if m.history.store != nil {
-							_ = m.history.store.Delete(key)
-						}
-						m.history.items = append(m.history.items[:i], m.history.items[i+1:]...)
-					}
-				}
-				items := make([]list.Item, len(m.history.items))
-				for i, it := range m.history.items {
-					it.isSelected = nil
-					it.isMarkedForDeletion = nil
-					m.history.items[i] = it
-					items[i] = it
-				}
-				m.history.list.SetItems(items)
-				if len(m.history.items) == 0 {
-					m.history.list.Select(-1)
-				} else if m.history.list.Index() >= len(m.history.items) {
-					m.history.list.Select(len(m.history.items) - 1)
-				}
-				m.history.selectionAnchor = -1
-			})
-			m.confirmCancel = func() {
-				for i := range m.history.items {
-					m.history.items[i].isMarkedForDeletion = nil
-				}
-			}
-		} else if m.ui.focusOrder[m.ui.focusIndex] == idTopics && m.topics.selected >= 0 && m.topics.selected < len(m.topics.items) {
-			idx := m.topics.selected
-			name := m.topics.items[idx].title
-			m.confirmReturnFocus = m.ui.focusOrder[m.ui.focusIndex]
-			m.startConfirm(fmt.Sprintf("Delete topic '%s'? [y/n]", name), "", func() {
-				m.removeTopic(idx)
-				if m.currentMode() == modeTopics {
-					m.rebuildActiveTopicList()
-				}
-			})
-		}
-	default:
-		switch msg.String() {
-		case "ctrl+b":
-			if err := m.connections.manager.LoadProfiles(""); err != nil {
-				m.appendHistory("", err.Error(), "log", err.Error())
-			}
-			m.refreshConnectionItems()
-			m.saveCurrent()
-			m.savePlannedTraces()
-			cmds = append(cmds, m.setMode(modeConnections))
-		case "ctrl+t":
-			m.topics.panes.subscribed = paneState{sel: 0, page: 0, index: 0, m: m}
-			m.topics.panes.unsubscribed = paneState{sel: 0, page: 0, index: 1, m: m}
-			m.topics.panes.active = 0
-			m.topics.list.SetSize(m.ui.width/2-2, m.ui.height-4)
-			m.rebuildActiveTopicList()
-			cmds = append(cmds, m.setMode(modeTopics))
-		case "ctrl+p":
-			items := []list.Item{}
-			for _, pld := range m.message.payloads {
-				items = append(items, payloadItem{topic: pld.topic, payload: pld.payload})
-			}
-			m.message.list = list.New(items, list.NewDefaultDelegate(), m.ui.width-4, m.ui.height-4)
-			m.message.list.DisableQuitKeybindings()
-			m.message.list.SetShowTitle(false)
-			cmds = append(cmds, m.setMode(modePayloads))
-		case "ctrl+r":
-			m.traces.list.SetSize(m.ui.width-4, m.ui.height-4)
-			cmds = append(cmds, m.setMode(modeTracer))
-		}
-	}
-	if len(cmds) == 0 {
-		return nil
-	}
-	return tea.Batch(cmds...)
 }
 
 // handleClientMouse processes mouse events in client mode.
